requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0,<4.0.0
python-docx>=0.8.11
pandas>=2.0.0
requests>=2.31.0
